Updated HTML, CSS, JS
HTML - Updated nav with navbar from bootstrap as well as changed the modal from bootstrap as well.
CSS - Updated the look of the navigation button as well as the modal and background of every element
JS - Implemented bootstrap class as well as the modal calling from bootstrap. Practiced a little jQuery to understand the difference syntax from vanilla JS and jQuery